requirements.txt
Flask==3.1.0
flask-cors==5.0.1
bcrypt==4.3.0
PyJWT==2.10.1
python-dotenv==1.1.0
mysql-connector-python==9.3.0
